+++ Changes.md
@@ -4,6 +4,7 @@
 Improvements
 ------------
 
+- Viewer : Added standard Top, Front and Side orthographic cameras, accessible from the Camera menu (#4523).
 - Spreadsheet : Improved drag and drop of lists of strings. <kbd>Shift</kbd> may now be held to add to the current list, and <kbd>Control</kbd> may be held to remove from the current list (#3894). This is particularly useful when a PathFilter has been added to a spreadsheet.
 
 Fixes